Add files via upload (#6)
added two data runs in the required format . Both are computed for the trappist1b evolution at 10**5 years. Data1 has the total surface pressure scaled to the hydrogen abundance predicted by proteus. Data2 has a not scaled total pressure, equal to the total pressure of CO2 and H2O in the evolution model.